review: cover founder-facing createGitHubAppClient + harden PEM tests (P1)
Multi-agent review (pattern-recognition + code-quality, concurring) found a
third new App() site — createGitHubAppClient in app-client.ts, the
FOUNDER-facing path — reads the same GITHUB_APP_PRIVATE_KEY raw and mints JWTs
via the identical vulnerable @octokit/app codepath. The original fix only
covered the synthetic-probe sites, leaving the higher-traffic founder path
exposed to the same decode failure.

- Extract normalizeAppPrivateKey into a neutral shared module
  server/github/app-private-key.ts; route all three new App() sites through it
  (probe-octokit x2 + app-client). Avoids the founder client depending on the
  synthetic-probe module (review P3).
- Correct the helper comment: on the Node runtime universal-github-app-jwt
  self-converts PKCS#1, so CRLF / escaped-\n line endings are the operative
  bug, not key format (review P3).
- Harden probe-octokit-pem.test.ts: AC2 now reassembles + parses the extracted
  body to prove intact DER (Buffer.from base64 is lenient); add typeof guards
  so the empty/whitespace throws-tests can't pass vacuously on a missing export
  (review P3).
- app-client.test.ts: replace bogus stub key with a synthesized keypair, since
  normalization now runs before the mocked App.
